# docs/openapi.yaml
openapi: 3.0.3
info:
  title: 回测平台 API
  version: 1.0.0
  description: |
    回测平台提供组合回测、资产分析、蒙特卡洛模拟、组合优化、有效前沿计算、
    战术分配、信号分析、PCA 主成分分析、LETF 滑点分析、目标优化等量化投资工具。

    ## 认证
    - 管理端点（admin、data/manage）需要 `x-api-key` 请求头认证
    - 计算密集型端点（backtest、backtest-optimizer）可选 `x-api-key`，有 Key 时验证身份，无 Key 时放行但记录匿名访问
    - 健康检查和 metrics 端点不需要认证

    ## 速率限制
    - 普通 API：100 次/15 分钟/IP
    - 计算密集型 API（backtest、backtest-optimizer）：10 次/分钟/IP
servers:
  - url: http://localhost:5001/api/v1
    description: 本地开发环境
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT 访问令牌，通过 /auth/login 获取
...
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          oneOf:
            - type: string
            - type: object
              required:
                - code
                - message
              properties:
                code:
                  type: string
                  description: 机器可读错误码
                message:
                  type: string
                  description: 人类可读错误描述

    # ===== 认证鉴权 =====
    # 企业理由：API 契约文档是前后端协作的唯一可靠依据。
    # 认证流程涉及令牌签发、刷新、撤销，任何字段歧义都会导致集成故障，
    # 因此必须将 LoginRequest / TokenResponse / UserInfo 等结构固化到 OpenAPI 规范中，
    # 确保客户端与服务端对令牌生命周期管理达成一致理解。
...
    CashflowLeg:
      type: object
      required:
        - id
        - amount
        - type
        - frequency
        - offset
      properties:
        id:
          type: string
        amount:
          type: number
        type:
          type: string
          enum: [contribution, withdrawal]
        frequency:
          type: string
          enum: [yearly, monthly, quarterly, weekly]
        offset:
          type: number
        until:
          type: string
          description: 结束日期
    OneTimeCashflow:
      type: object
      required:
        - id
        - amount
        - type
        - date
      properties:
        id:
          type: string
        amount:
          type: number
        type:
          type: string
          enum: [contribution, withdrawal]
        date:
          type: string
          format: date
    BacktestParameters:
      type: object
      required:
        - startDate
        - endDate
        - startingValue
        - adjustForInflation
        - rollingWindowMonths
        - benchmarkTicker
      properties:
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        startingValue:
          type: number
        baseCurrency:
          type: string
          enum: [usd, cny]
          default: usd
        adjustForInflation:
          type: boolean
        rollingWindowMonths:
          type: number
          default: 12
        benchmarkTicker:
          type: string
        extendedWithdrawalStats:
          type: boolean
        cashflowLegs:
          type: array
          items:
            $ref: '#/components/schemas/CashflowLeg'
        oneTimeCashflows:
          type: array
          items:
            $ref: '#/components/schemas/OneTimeCashflow'

    # ===== 统计 =====
...
    OptimizeRequest:
      type: object
      required:
        - tickers
        - parameters
      properties:
        tickers:
          type: array
          items:
            type: string
          maxItems: 50
        objective:
          type: string
          enum: [maxSharpe, minVolatility, maxReturn]
          description: 优化目标
        constraints:
          type: object
          properties:
            minWeight:
              type: number
            maxWeight:
              type: number
        parameters:
          $ref: '#/components/schemas/BacktestParameters'
        riskFreeRate:
          type: number
        numIterations:
          type: number
          description: 迭代次数，上限 100000，默认 10000
    BacktestOptimizerRequest:
      type: object
      required:
        - portfolio
        - parameterSpace
        - parameters
        - objective
      properties:
        portfolio:
          type: object
          required:
            - assets
          properties:
            name:
              type: string
            assets:
              type: array
              items:
                type: object
                required:
                  - ticker
                  - weight
                properties:
                  ticker:
                    type: string
                  weight:
                    type: number
        parameterSpace:
          type: object
          required:
            - rebalanceFrequencies
            - initialCapital
          properties:
            rebalanceFrequencies:
              type: array
              items:
                type: string
                enum: [daily, weekly, monthly, quarterly, annual, none, threshold]
            rebalanceThreshold:
              type: object
              properties:
                min:
                  type: number
                max:
                  type: number
                step:
                  type: number
            initialCapital:
              type: object
              required:
                - min
                - max
                - step
              properties:
                min:
                  type: number
                max:
                  type: number
                step:
                  type: number
        parameters:
          type: object
          required:
            - startDate
            - endDate
          properties:
            startDate:
              type: string
              format: date
            endDate:
              type: string
              format: date
            benchmarkTicker:
              type: string
            baseCurrency:
              type: string
              enum: [usd, cny]
            adjustForInflation:
              type: boolean
        objective:
          type: string
          enum: [maxCagr, minMaxDrawdown, maxSharpe, maxSortino]
        constraints:
          type: object
          properties:
            maxDrawdown:
              type: number
              description: 最大回撤百分比，如 20 表示 20%
            minCagr:
              type: number
              description: 最低年化收益百分比

    # ===== 战术网格搜索 =====
...
paths:
...
  /backtest/optimize:
    post:
      tags: [backtest]
      summary: 运行组合优化
      description: 根据目标函数和约束条件优化资产权重
      security:
        - ApiKeyAuth: []
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OptimizeRequest'
      responses:
        '200':
          description: 优化结果
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 参数校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 优化运行失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /backtest-optimizer/optimize:
    post:
      tags: [backtest-optimizer]
      summary: 回测参数优化
      description: |
        遍历参数空间（再平衡频率 × 阈值 × 初始资金）运行组合回测，
        按优化目标排序返回全部结果及最优组合的收益曲线。
        参数组合上限 1000。
      security:
        - ApiKeyAuth: []
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BacktestOptimizerRequest'
      responses:
        '200':
          description: 优化结果
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      results:
                        type: array
                        items:
                          type: object
                          properties:
                            rebalanceFrequency:
                              type: string
                            rebalanceThreshold:
                              type: number
                            initialCapital:
                              type: number
                            cagr:
                              type: number
                            maxDrawdown:
                              type: number
                            sharpe:
                              type: number
                            sortino:
                              type: number
                            stdev:
                              type: number
                            calmar:
                              type: number
                      best:
                        type: object
                        nullable: true
                        description: 最优组合（含收益曲线）
                      benchmarkGrowth:
                        type: array
                        nullable: true
                        items:
                          type: object
                          properties:
                            date:
                              type: string
                            value:
                              type: number
                      totalCombinations:
                        type: number
        '400':
          description: 参数校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 优化运行失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'